StretchController
Keeps track of when to release a grip because of to much distance
between the hands. It releases the oldest gripping hand. A break limit
is set on the script, this value can be modified to change when the
hands are supposed to break ("release").
The script can be found on the frogs 'body' transform.